[DefaultCanvas] Use via width for spg [DefaultCanvas] Split into separate file
